Add method MVWorldManager#addOrRemoveWorldSafely
Addresses Multiverse/Multiverse-Core#2560

Commands that load or unload worlds trigger an IllegalStateException if they are run via a command block while the worlds are being ticked.

Using a BukkitRunnable, the operation can be delayed until the next tick at a time when the worlds are not being ticked.

MVWorldManager#addOrRemoveWorldSafely performs this logic, either running the operation now or delaying it.

8 commands were modified to use MVWorldManager#addOrRemoveWorldSafely, which I think are all the relevant commands.

Unfortunately I haven't found a way to tell when the worlds are being ticked on a Spigot server. There probably won't be any way to do this until [SPIGOT-7089](https://hub.spigotmc.org/jira/browse/SPIGOT-7089) resolves